Refactored internals a bit to simplify things
