--- src/ReactGraph.Tests/BasicType.cs
@@ -1,9 +1,0 @@
-namespace ReactGraph.Tests
-{
-    public class BasicType
-    {
-        public string Target { get; set; }
-        public int Source1 { get; set; }
-        public int Source2 { get; set; }
-    }
-}
